:art: Rename: app to labsys and main to samples.
