+++ labsys/__init__.py
@@ -32,7 +32,7 @@ def create_app(config_name):
         from flask_sslify import SSLify
         sslify = SSLify(app)
 
-    from .main import main as main_blueprint
+    from .samples import main as main_blueprint
     app.register_blueprint(main_blueprint)
 
     from .auth import auth as auth_blueprint
