+++ labsys/inventory/models.py
@@ -12,7 +12,7 @@ from sqlalchemy import asc
 
 from flask import current_app
 
-from app import db, login_manager
+from labsys import db, login_manager
 
 
 class Product(db.Model):
