+++ labsys/inventory/views.py
@@ -2,11 +2,11 @@ from flask import render_template, redirect, url_for, flash, abort, request
 from flask_login import current_user, login_required
 
 from .utils import stock_is_at_minimum, export_table
-from app.decorators import permission_required
-from app.email import send_email
-from app import db
-from app.auth.models import Permission, User
-from app.inventory.models import Transaction, Product, StockProduct
+from labsys.decorators import permission_required
+from labsys.email import send_email
+from labsys import db
+from labsys.auth.models import Permission, User
+from labsys.inventory.models import Transaction, Product, StockProduct
 from . import inventory
 from .forms import AddTransactionForm, SubTransactionForm, ProductForm
