+++ labsys/templates/auth/email/confirm.html
