+++ labsys/templates/auth/email/confirm.txt
