+++ labsys/templates/inventory/_stock.html
