+++ labsys/templates/inventory/create-reactive.html
