+++ labsys/templates/inventory/create-transaction.html
