+++ labsys/templates/inventory/email/stock_alert.txt
