+++ labsys/templates/inventory/index.html
