+++ labsys/templates/inventory/list-catalog.html
