+++ labsys/templates/inventory/list-transactions.html
