+++ labsys/templates/mail/new_user.txt
